Refactor code to add German negation
