+++ negate/__init__.py
@@ -2,5 +2,5 @@ from .negate import Negator
 from .tokens import Token
 
 # Don't expose the following submodules.
-del globals()["negate"]
-del globals()["tokens"]
+#del globals()["negate"]
+#del globals()["tokens"]
